requirements.txt
torch
Whisper~=1.1.10
git+https://github.com/openai/whisper.git
numpy==2.0.2
python-telegram-bot
requests~=2.32.3
python-dotenv~=1.0.1
telegram~=0.0.1